further implementation of edit
