+++ package.json
@@ -12,16 +12,16 @@
     "google-maps-react": "^2.0.6",
     "leaflet": "^1.7.1",
     "leaflet-geosearch": "^3.3.2",
+    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "prop-types": "^15.7.2",
     "react-dropzone": "^11.3.2",
     "react-google-maps": "^9.4.5",
     "react-hook-form": "^7.1.0",
     "react-hotkeys-hook": "^3.3.1",
+    "react-icons": "^4.2.0",
     "react-leaflet": "^3.1.0",
     "react-places-autocomplete": "^7.3.0",
-    "react-icons": "^4.2.0",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.2",
@@ -36,6 +36,7 @@
     "sockjs-client": "^1.4.0",
     "stompjs": "^2.3.3",
     "styled-components": "^5.2.1",
+    "typescript": "^4.3.2",
     "validator": "^13.6.0"
   },
   "scripts": {
