--- src/components/HomeScreen/landy-react-template/common/Button/index.tsx
@@ -1,13 +1,0 @@
-import { StyledButton } from "./styles";
-import { ButtonProps } from "../types";
-
-export const Button = ({
-  color,
-  fixedWidth,
-  children,
-  onClick,
-}: ButtonProps) => (
-  <StyledButton color={color} fixedWidth={fixedWidth} onClick={onClick}>
-    {children}
-  </StyledButton>
-);
