--- src/components/HomeScreen/landy-react-template/common/Container/index.tsx
@@ -1,8 +1,0 @@
-import { StyledContainer } from "./styles";
-import { ContainerProps } from "../types";
-
-const Container = ({ border, children }: ContainerProps) => (
-  <StyledContainer border={border}>{children}</StyledContainer>
-);
-
-export default Container;
