--- src/components/HomeScreen/landy-react-template/common/SvgIcon/index.tsx
@@ -1,5 +1,0 @@
-import { SvgIconProps } from "../types";
-
-export const SvgIcon = ({ src, width, height }: SvgIconProps) => (
-  <img src={`/img/svg/${src}`} alt={src} width={width} height={height} />
-);
