--- src/components/HomeScreen/landy-react-template/common/utils/validationRules.ts
@@ -1,18 +1,0 @@
-import { validateProps } from "../types";
-
-export default function validate(values: validateProps) {
-  let errors = {} as validateProps;
-
-  if (!values.name) {
-    errors.name = "Name is required";
-  }
-  if (!values.email) {
-    errors.email = "Email address is required";
-  } else if (!/\S+@\S+\.\S+/.test(values.email)) {
-    errors.email = "Email address is invalid";
-  }
-  if (!values.message) {
-    errors.message = "Message is required";
-  }
-  return errors;
-}
