--- src/components/HomeScreen/landy-react-template/components/ContactForm/types.ts
@@ -1,10 +1,0 @@
-export interface ContactProps {
-  title: string;
-  content: string;
-  id: string;
-  t: any;
-}
-
-export interface ValidationTypeProps {
-  type: any;
-}
