--- src/components/HomeScreen/landy-react-template/components/ContentBlock/index.tsx
@@ -1,11 +1,0 @@
-import LeftContentBlock from "./LeftContentBlock";
-import RightContentBlock from "./RightContentBlock";
-import { ContentBlockProps } from "./types";
-
-const ContentBlock = (props: ContentBlockProps) => {
-  if (props.type === "left") return <LeftContentBlock {...props} />;
-  if (props.type === "right") return <RightContentBlock {...props} />;
-  return null;
-};
-
-export default ContentBlock;
