--- src/components/HomeScreen/landy-react-template/components/ContentBlock/types.ts
@@ -1,10 +1,0 @@
-export interface ContentBlockProps {
-  icon: string;
-  title: string;
-  content: string;
-  section?: any;
-  button?: any;
-  t?: any;
-  id: string;
-  type?: string;
-}
