--- src/components/HomeScreen/landy-react-template/components/MiddleBlock/index.tsx
@@ -1,42 +1,0 @@
-import { Row, Col } from "antd";
-import { withTranslation } from "react-i18next";
-import { Slide } from "react-awesome-reveal";
-import { Button } from "../../common/Button";
-import { MiddleBlockSection, Content, ContentWrapper } from "./styles";
-
-interface MiddleBlockProps {
-  title: string;
-  content: string;
-  button: string;
-  t: any;
-}
-
-const MiddleBlock = ({ title, content, button, t }: MiddleBlockProps) => {
-  const scrollTo = (id: string) => {
-    const element = document.getElementById(id) as HTMLDivElement;
-    element.scrollIntoView({
-      behavior: "smooth",
-    });
-  };
-  return (
-    <MiddleBlockSection>
-      <Slide direction="up">
-        <Row justify="center" align="middle">
-          <ContentWrapper>
-            <Col lg={24} md={24} sm={24} xs={24}>
-              <h6>{t(title)}</h6>
-              <Content>{t(content)}</Content>
-              {button && (
-                <Button name="submit" onClick={() => scrollTo("mission")}>
-                  {t(button)}
-                </Button>
-              )}
-            </Col>
-          </ContentWrapper>
-        </Row>
-      </Slide>
-    </MiddleBlockSection>
-  );
-};
-
-export default withTranslation()(MiddleBlock);
