--- src/components/HomeScreen/landy-react-template/content/ContactContent.json
@@ -1,4 +1,0 @@
-{
-    "title": "Contact Form",
-    "text": "The following form demonstrates form validation in action. Contact form component reduces the amount of time it is being re-rendered by the user as it embraces uncontrolled form validation to reduce any unnecessary performance penalty."
-}
