--- src/components/HomeScreen/landy-react-template/content/IntroContent.json
@@ -1,13 +1,0 @@
-{
-  "title": "Landing page template for developers & startups",
-  "text": "Beautifully designed templates using React.js, ant design and styled-components! Save weeks of time and build your landing page in minutes.",
-  "button": [
-    {
-      "title": "Explore"
-    },
-    {
-      "title": "Learn More",
-      "color": "#fff"
-    }
-  ]
-}
