--- src/components/HomeScreen/landy-react-template/content/MiddleBlockContent.json
@@ -1,6 +1,0 @@
-{
-    "title": "Introduce Your Product Quickly & Effectively",
-    "text": "Landy comes with multi-lingual support, all of the content is stored in the JSON files, so that you can manage the texts without having any prior knowledge in developing web applications using React.js.",
-    "button": "Get Started"
-  }
-
