--- src/components/HomeScreen/landy-react-template/content/MissionContent.json
@@ -1,4 +1,0 @@
-{
-  "title": "Ready made sections",
-  "text": "Landy comes with a set of ready to use sections to help you easily create a landing page for your own brand, with all the content your startup desires. In order to replace any dummy elements (text, image, etc..) all you need to do is go to content folder and input your real content."
-}
