--- src/components/HomeScreen/landy-react-template/content/ProductContent.json
@@ -1,4 +1,0 @@
-{
-  "title": "That's about it!",
-  "text": "If you are familiar with React, or SPA frameworks in general, there shouldn't be any confusing parts about this template. if not, you can always reach out to me and I'll try me best to answer your questions."
-}
