--- src/components/HomeScreen/landy-react-template/locales/en/translation.json
@@ -1,28 +1,0 @@
-{
-  "About": "About",
-  "Policy": "Policy",
-  "Mission": "Mission",
-  "Product": "Product",
-  "Tell us everything": "Tell us everything",
-  "Contact": "Contact",
-  "Do you have any question? Feel free to reach out.": "Do you have any question? Feel free to reach out.",
-  "Let's Chat": "Let's Chat",
-  "Application Security": "Application Security",
-  "Software Principles": "Software Principles",
-  "Support Center": "Support Center",
-  "Customer Support": "Customer Support",
-  "ADDRESS": "ADDRESS",
-  "Press": "Press",
-  "Careers & Culture": "Careers & Culture",
-  "Language": "Language",
-  "Explore": "Explore",
-  "Learn More": "Learn More",
-  "Get Started": "Get Started",
-  "Submit": "Submit",
-  "Message" : "Message",
-  "Your Message": "Your Message",
-  "Name" : "Name",
-  "Your Name": "Your Name",
-  "Email" : "Email",
-  "Your Email": "Your Email"
-}
