--- src/components/HomeScreen/landy-react-template/locales/es/translation.json
@@ -1,28 +1,0 @@
-{
-  "About": "Acerca de",
-  "Policy": "POLÍTICA",
-  "Mission": "Misión",
-  "Product": "Producto",
-  "Tell us everything": "Cuéntanos todo",
-  "Contact": "Contacto",
-  "Do you have any question? Feel free to reach out.": "¿Tienes alguna pregunta? No dude en contactarnos",
-  "Let's Chat": "Charlemos",
-  "Application Security": "Seguridad de la aplicación",
-  "Software Principles": "Principios de software",
-  "Support Center": "Centro de Apoyo",
-  "Customer Support": "Atención al cliente",
-  "ADDRESS": "HABLA A",
-  "Press": "Prensa",
-  "Careers & Culture": "Carreras & cultura",
-  "Language": "Language",
-  "Explore": "Explorar",
-  "Learn More": "Más información",
-  "Get Started": "Comenzar",
-  "Submit": "Enviar",
-  "Message" : "Mensaje",
-  "Your Message": "Tu Mensaje",
-  "Name" : "Nombre",
-  "Your Name": "Tu nombre",
-  "Email" : "Email",
-  "Your Email": "Tu correo electrónico"
-}
