+++ src/components/profilepage/Edit.js
@@ -2,8 +2,11 @@ import React, {useState} from "react";
 import {Modal, ButtonToolbar, Button} from "rsuite";
 import {TextField} from "@material-ui/core";
 import {api, handleError} from "../../helpers/api";
+import {useJsApiLoader} from "@react-google-maps/api";
+import PlacesAutocomplete, {geocodeByAddress, getLatLng} from "react-places-autocomplete";
+require('dotenv').config();
 
-//TODO: change the edit modal to fetch latLng from LocationSearchInput
+const API_KEY = process.env.REACT_APP_GOOGLE_MAPS_API_KEY;
 
 export default function Edit(props) {
     const id = localStorage.getItem("id")
@@ -17,6 +20,30 @@ export default function Edit(props) {
         postcode: props.userdata.postcode,
     });
 
+    const {isLoaded} = useJsApiLoader({
+        id: 'google-map-script',
+        googleMapsApiKey: API_KEY,
+    });
+
+    const [address, setAddress] = useState();
+    const [newLocation, setNewLocation] = useState();
+
+    function handleInput(val) {
+        setAddress(val);
+    }
+
+    function handleSelect(val) {
+        setAddress(val);
+        geocodeByAddress(address)
+          .then(results => {
+              getLatLng(results[0])
+              console.log(results)
+          })
+          .then(latLng => setNewLocation(latLng))
+          .catch(error => console.error('Error', error));
+        console.log(newLocation);
+        // api call set lng lat
+    }
     const handleChange = (e) => {
         const {id, value} = e.target
         setEdit((prevState) => ({...prevState, [id]: value}));
@@ -37,7 +64,6 @@ export default function Edit(props) {
             await api.put('/users/' + id, requestBody);
             window.location.reload(false);
             close()
-
         } catch (error) {
             alert(`Something went wrong during the login: \n${handleError(error)}`);
         }
@@ -51,7 +77,7 @@ export default function Edit(props) {
         setModal({show: true});
     }
 
-    return (
+    return isLoaded ? (
         <div className="modal-container">
             <ButtonToolbar>
                 <Button onClick={open}> Edit Profile</Button>
@@ -96,6 +122,50 @@ export default function Edit(props) {
                       id="password"
                       onChange={handleChange}
                     />
+                    <PlacesAutocomplete
+                      value={address}
+                      onChange={handleInput}
+                      onSelect={handleSelect}
+                    >
+                        {({getInputProps, suggestions, getSuggestionItemProps, loading}) => (
+                          <div>
+                              <TextField
+                                {...getInputProps({
+                                    placeholder: 'Search Places ...',
+                                    className: 'location-search-input',
+                                })}
+                                id="outlined-basic"
+                                margin="normal"
+                                label="City/Street"
+                                variant="outlined"
+                                fullWidth
+                                required
+                              />
+                              <div className="autocomplete-dropdown-container">
+                                  {loading && <div>Loading...</div>}
+                                  {suggestions.map(suggestion => {
+                                      const className = suggestion.active
+                                        ? 'suggestion-item--active'
+                                        : 'suggestion-item';
+                                      // inline style for demonstration purpose
+                                      const style = suggestion.active
+                                        ? {backgroundColor: '#fafafa', cursor: 'pointer'}
+                                        : {backgroundColor: '#ffffff', cursor: 'pointer'};
+                                      return (
+                                        <div
+                                          {...getSuggestionItemProps(suggestion, {
+                                              className,
+                                              style,
+                                          })}
+                                        >
+                                            <span>{suggestion.description}</span>
+                                        </div>
+                                      );
+                                  })}
+                              </div>
+                          </div>
+                        )}
+                    </PlacesAutocomplete>
                 </Modal.Body>
                 <Modal.Footer>
                     <Button
@@ -117,5 +187,5 @@ export default function Edit(props) {
                 </Modal.Footer>
             </Modal>
         </div>
-        );
+        ) : <></> ;
 }
